--- a/app/static/index.html
+++ b/app/static/index.html
@@ -80,6 +80,7 @@
     </div>
     
     <script src="/static/js/progress.js" defer></script>
+    <script src="/static/js/editScreen.js" defer></script>
     <script src="/static/js/recorder.js" defer></script>
     <script src="/static/js/ui.js" defer></script>
 </body>