… tests/api/test_draft_management.py
… tests/frontend/test_edit_screen.py
… workflow/logs/development.jsonl