# Development Guidance: Issue #18 - Create Review/Edit Screen

## Implementation Overview

This development guide outlines the implementation of a mobile-friendly review/edit screen for voice transcriptions, following the established FastAPI + Vanilla JS architecture.

## Branch Strategy
- **Feature Branch**: `feature/issue-18-review-edit-screen`
- **Base Branch**: `feature/issue-1-docker-fastapi-setup`
- **Commit Pattern**: Conventional commits (`feat:`, `fix:`, `test:`)

## Architecture Alignment

### Frontend Implementation
Following the existing Vanilla JS + Web APIs pattern:
- **New Module**: `app/static/js/editScreen.js` - Edit interface controller
- **Styling**: Extend `app/static/css/main.css` with edit-specific styles
- **Integration**: Modify existing `recorder.js` to handle edit workflow

### Backend Implementation
Extending the existing session management system:
- **New Models**: Draft management models in `app/models/session.py`
- **API Endpoints**: Draft management in `app/api/sessions.py`
- **Services**: Extend session manager for draft persistence

## Technical Implementation Plan

### Phase 1: Data Models & Backend API

#### 1.1 Extend Session Models
```python
# app/models/session.py - Add draft models
class DraftData(BaseModel):
    transcription: Optional[str] = None
    summary: Optional[List[str]] = None
    keywords: Optional[List[str]] = None
    last_modified: datetime = Field(default_factory=datetime.utcnow)

class SessionUpdateRequest(BaseModel):
    # Extend existing model
    draft: Optional[DraftData] = None
```

#### 1.2 Session API Extensions
```python
# app/api/sessions.py - Add draft endpoints
@router.patch("/{session_id}/draft")
async def update_draft(
    session_id: str,
    draft: DraftData,
    request: Request
) -> SessionResponse:
    """Update session draft data"""

@router.get("/{session_id}/preview")
async def preview_markdown(
    session_id: str,
    request: Request
) -> dict:
    """Generate markdown preview from current draft"""
```

### Phase 2: Frontend Edit Interface

#### 2.1 Edit Screen Module
```javascript
// app/static/js/editScreen.js
class EditScreen {
    constructor(sessionId, sessionData) {
        this.sessionId = sessionId;
        this.transcription = sessionData.transcription;
        this.summary = sessionData.summary || [];
        this.keywords = sessionData.keywords || [];
        this.autoSaveTimer = null;
        this.undoStack = [];
        this.redoStack = [];
    }

    async render() {
        // Create edit UI with responsive design
        // Implement transcription editor
        // Add summary bullet editor
        // Create keyword tag manager
        // Setup auto-save and validation
    }

    setupAutoSave() {
        // Save draft every 10 seconds
    }

    handleUndo() {
        // Implement undo/redo functionality
    }

    async saveToVault() {
        // Final save workflow integration
    }
}
```

#### 2.2 UI Integration
- Modify `recorder.js` to transition to edit screen after transcription
- Update `ui.js` to handle edit screen navigation
- Extend `main.css` with mobile-first edit styles

### Phase 3: Mobile UX Optimization

#### 3.1 Touch-Friendly Design
- 44px minimum touch targets
- Smooth virtual keyboard handling
- Swipe gestures for section navigation
- Debounced input handling

#### 3.2 Session Persistence
- localStorage backup for edits
- Automatic recovery on page refresh
- Warning system for unsaved changes

## File Structure Changes

```
app/static/
├── css/
│   └── main.css           # Extended with edit styles
├── js/
│   ├── editScreen.js      # New: Edit interface controller
│   ├── recorder.js        # Modified: Add edit workflow
│   ├── ui.js              # Modified: Edit navigation
│   └── progress.js        # Existing: Reused for save progress
└── index.html             # Modified: Add edit screen sections

app/models/
└── session.py             # Extended: Draft data models

app/api/
└── sessions.py            # Extended: Draft management endpoints

app/services/
└── session_manager.py     # Extended: Draft persistence logic
```

## Testing Strategy

### Unit Tests
```python
# tests/api/test_sessions.py - Extend existing tests
async def test_update_draft():
    """Test draft update endpoint"""

async def test_preview_markdown():
    """Test markdown preview generation"""

# tests/services/test_session_manager.py - Extend existing tests
async def test_draft_persistence():
    """Test draft data persistence"""
```

### Frontend Tests
```python
# tests/frontend/test_edit_screen.py
def test_edit_screen_rendering():
    """Test edit interface creation"""

def test_auto_save_functionality():
    """Test auto-save timer and persistence"""

def test_undo_redo_operations():
    """Test undo/redo functionality"""
```

### Integration Tests
```python
# tests/integration/test_edit_workflow.py
async def test_full_edit_workflow():
    """Test record → edit → save workflow"""

async def test_session_recovery():
    """Test edit session persistence across refreshes"""
```

## Quality Requirements

### Code Standards
- **Python**: Black formatting, mypy type hints, 80%+ coverage
- **JavaScript**: 2-space indentation, const/let (no var), ESLint compliance
- **CSS**: Mobile-first responsive design, WCAG 2.1 AA accessibility

### Performance Targets
- Edit screen load: <1s after transcription
- Input responsiveness: <100ms lag
- Save operation: <2s completion
- Auto-save frequency: Every 10 seconds

### Accessibility
- ARIA labels for screen readers
- Keyboard navigation support
- High contrast text (minimum 4.5:1 ratio)
- Touch target minimum 44px

## Implementation Steps

### Step 1: Backend Foundation (Day 1)
1. Extend session models with draft data
2. Add draft management endpoints
3. Update session manager for draft persistence
4. Write unit tests for new backend functionality

### Step 2: Edit Interface Core (Day 2-3)
1. Create `editScreen.js` module
2. Implement transcription text editor
3. Add summary bullet-point manager
4. Create keyword tag interface
5. Setup auto-save functionality

### Step 3: UX Enhancement (Day 4)
1. Add undo/redo functionality
2. Implement markdown preview
3. Setup session persistence with localStorage
4. Add validation and character counting

### Step 4: Integration & Polish (Day 5)
1. Integrate with existing recorder workflow
2. Add mobile gesture optimization
3. Implement accessibility features
4. Complete integration testing

## Risk Mitigation

### Mobile Text Editing Complexity
- **Strategy**: Progressive enhancement from basic textarea
- **Testing**: Extensive mobile device testing
- **Fallback**: Simplified interface for older browsers

### Session Data Loss
- **Strategy**: Dual persistence (localStorage + server)
- **Implementation**: Auto-save every 10 seconds
- **Recovery**: Automatic draft restoration on page load

### Performance Issues
- **Strategy**: Debounced input handling
- **Monitoring**: Performance benchmarks in tests
- **Optimization**: Virtual scrolling for large transcriptions

## Pull Request Checklist

### Before Creating PR
- [ ] All unit tests pass (`pytest tests/ -v`)
- [ ] Code formatted (`black app tests && isort app tests`)
- [ ] Type checking passes (`mypy app`)
- [ ] Integration tests pass
- [ ] Mobile testing completed on 3+ devices
- [ ] Accessibility audit passed
- [ ] Performance benchmarks met

### PR Content
- [ ] Clear title: "feat: implement review/edit screen for voice transcriptions"
- [ ] Comprehensive description linking to issue #18
- [ ] Screenshots/video of mobile interface
- [ ] Performance metrics included
- [ ] Breaking changes documented (if any)

## Documentation Updates

### User Documentation
- Update README with edit workflow screenshots
- Add troubleshooting section for mobile editing issues

### Developer Documentation
- Document new API endpoints in OpenAPI spec
- Add frontend architecture notes for edit system
- Update testing guide with edit workflow examples

## Definition of Done Verification

### Functional Requirements
- [ ] Edit transcription with mobile-friendly interface
- [ ] Modify summary bullet points
- [ ] Add/remove keywords with tag interface
- [ ] Save/discard workflow with confirmation
- [ ] Session persistence across browser events

### Technical Requirements
- [ ] <3 taps from edit to save (UX requirement)
- [ ] <1s edit screen load time
- [ ] <2s final save operation
- [ ] 99.9% session persistence success rate
- [ ] WCAG 2.1 AA accessibility compliance

### Quality Gates
- [ ] 80%+ test coverage for new code
- [ ] All linting and type checking passes
- [ ] Mobile responsive design verified
- [ ] Integration with existing workflow confirmed
- [ ] Performance benchmarks met

---

*This development plan ensures seamless integration with the existing Dialtone architecture while delivering the review/edit functionality required for Sprint 3.*